moved test driver to its own file
